move computeGroupIds to flowchart as its flowchart specific
